+++ firmware/SatStepperBegin/timers.h
@@ -10,7 +10,7 @@
 #ifndef _CONTROL_TIMER_H_
 #define _CONTROL_TIMER_H_
 
-#include <PeripheralHeaderIncludes.h>
+#include "stdint.h"
 
 typedef void (* _controlTimerInterruptHandler)(void);
 
@@ -18,9 +18,9 @@ void timer0Init(_controlTimerInterruptHandler);
 void timer1Init(_controlTimerInterruptHandler);
 
 // [period] = [10^3] / [System Clock Frequency in Hz]
-void setTimerPeriodByNum(unsigned timerNum, unsigned period);
-void stopTimerByNum(unsigned timerNum);
-void startTimerByNum(unsigned timerNum);
+void setTimerPeriodByNum(uint_fast8_t timerNum, uint_fast8_t period);
+void stopTimerByNum(uint_fast8_t timerNum);
+void startTimerByNum(uint_fast8_t timerNum);
 
 #endif //_CONTROL_TIMER_H_
